+++ tests/Endpoint/EnvelopeDocumentsEndpointTest.php
@@ -43,6 +43,12 @@ class EnvelopeDocumentsEndpointTest extends EndpointTestCase
         self::assertLastRequest('GET', '/api/envelopes/bar/documents/foo/tags?foo=bar');
     }
 
+    public function testTagsByAutoplacement(): void
+    {
+        self::endpoint()->createTagsByAutoplacement('foo');
+        self::assertLastRequest('POST', '/api/envelopes/bar/documents/foo/tags/by-autoplacement');
+    }
+
     public function testReplaceFile(): void
     {
         self::endpoint()->replaceFile('foo', ['foo' => 'bar']);