+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes will be documented in this file.
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
 ## [Unreleased]
+- Extend `EnvelopeRecipient` and `EnvelopeTemplateRecipient` resource
+- Add `EnvelopeDocument.createTagsByAutoplacement` endpoint
 - Add `Webhook oAuth properties`
 - Add `ForbiddenException`
 - Add `AccountBilling.productType`